Update the deployment script

+++ deploy.sh
@@ -1,27 +1,26 @@
 ### deploy all dot files
 
-rm ~/.zshrc
+[ -f ~/.zshrc ] && rm ~/.zshrc
 ln -s ~/dev/dotfiles/.zshrc ~/.zshrc
-rm ~/.myaliases.sh
+[ -f ~/.myaliases.sh ] && rm ~/.myaliases.sh
 ln -s ~/dev/dotfiles/.myaliases.sh ~/.myaliases.sh
-rm ~/.emacs
+[ -f ~/.emacs ] && rm ~/.emacs
 ln -s ~/dev/dotfiles/.emacs ~/.emacs
-rm ~/.gitconfig
+[ -f ~/.gitconfig ] && rm ~/.gitconfig
 ln -s ~/dev/dotfiles/.gitconfig ~/.gitconfig
 
 ##kitty the terminal emulator
-rm ~/.config/kitty/kitty.conf
+[ -f ~/.config/kitty/kitty.conf ] && rm ~/.config/kitty/kitty.conf
 ln -s ~/dev/dotfiles/kitty.conf ~/.config/kitty/kitty.conf
 
-rm ~/.Rprofile
+[ -f ~/.Rprofile ] && rm ~/.Rprofile
 ln -s ~/dev/dotfiles/.Rprofile ~/.Rprofile
 
 # rm ~/.hyper.js
 # ln -s ~/dev/dotfiles/.hyper.js ~/.hyper.js
 
-rm ~/.stumpwm
-ln -s ~/dev/dotfiles/stumpwm.lisp ~/.stumpwm
+[ -f ~/.stumpwmrc ] && rm ~/.stumpwmrc
+ln -s ~/dev/dotfiles/stumpwm.lisp ~/.stumpwmrc
 
 [ -f ~/.localrc ] && rm ~/.localrc
 [ -f ~/dev/dotfiles/.localrc ] && ln -s ~/dev/dotfiles/.localrc ~/.localrc
-
